config/requirements/dev.txt
# Local development dependencies

-r base.txt

django-debug-toolbar==1.8
django-celery==3.2.2
